code review: renaming
