+++ config/schemas/template.json
@@ -44,9 +44,9 @@
 				"treasureLikeZone" : { "type" : "number" },
 				"customObjectsLikeZone" : { "type" : "number" },
 
-				"visPositionX" : { "type" : "number" },
-				"visPositionY" : { "type" : "number" },
-				"visSize" : { "type" : "number" },
+				"visiblePositionX" : { "type" : "number" },
+				"visiblePositionY" : { "type" : "number" },
+				"visibleSize" : { "type" : "number" },
 				
 				"terrainTypes": {"$ref" : "#/definitions/stringArray"},
 				"bannedTerrains": {"$ref" : "#/definitions/stringArray"},
